common: move pxRadioPanel to the main project
