+++ pcsx2/gui/AppCommon.h
@@ -20,7 +20,7 @@
 #include "common/PersistentThread.h"
 
 #include "common/wxGuiTools.h"
-#include "common/pxRadioPanel.h"
+#include "gui/pxRadioPanel.h"
 #include "gui/pxCheckBox.h"
 #include "common/pxStaticText.h"
 #include "gui/CheckedStaticBox.h"
